refactor: extract table cell components into separate files for better organization + hide clientID
