+++ src/features/dashboard/sandboxes/table-config.tsx
@@ -2,27 +2,27 @@
 
 'use client'
 
-import { ArrowUpRight, Cpu } from 'lucide-react'
 import { ColumnDef, FilterFn, useReactTable } from '@tanstack/react-table'
 import { rankItem } from '@tanstack/match-sorter-utils'
-import { Sandbox, Template } from '@/types/api'
-import { Badge } from '@/ui/primitives/badge'
-import { PROTECTED_URLS } from '@/configs/urls'
+import { Sandbox } from '@/types/api'
 import { DateRange } from 'react-day-picker'
 import { isWithinInterval } from 'date-fns'
-import { CgSmartphoneRam } from 'react-icons/cg'
-import { cn } from '@/lib/utils'
-import { memo, useMemo } from 'react'
-import { Button } from '@/ui/primitives/button'
-import { useRouter } from 'next/navigation'
-import { useTemplateTableStore } from '../templates/stores/table-store'
-import { useServerContext } from '@/lib/hooks/use-server-context'
-import { JsonPopover } from '@/ui/json-popover'
+
+import {
+  CpuUsageCell,
+  RamUsageCell,
+  IdCell,
+  TemplateCell,
+  MetadataCell,
+  StartedAtCell,
+} from './table-cells'
 import posthog from 'posthog-js'
 import { logError } from '@/lib/clients/logger'
 import { ClientSandboxMetric } from '@/types/sandboxes.types'
 
-export type SandboxWithMetrics = Sandbox & { metrics: ClientSandboxMetric }
+export type SandboxWithMetrics = Sandbox & {
+  metrics?: ClientSandboxMetric | null
+}
 export type SandboxesTable = ReturnType<
   typeof useReactTable<SandboxWithMetrics>
 >
@@ -118,36 +118,11 @@ export const resourceRangeFilter: FilterFn<SandboxWithMetrics> = (
 export const fallbackData: SandboxWithMetrics[] = []
 
 export const COLUMNS: ColumnDef<SandboxWithMetrics>[] = [
-  // FIXME: Currently disabled due to issues with url state management when sandboxes dissapear
-  /*   {
-    id: 'pin',
-    cell: ({ row }) => (
-      <Button
-        variant="ghost"
-        size="icon"
-        className="text-fg-500 size-5"
-        onClick={() => row.pin(row.getIsPinned() ? false : 'top')}
-      >
-        {row.getIsPinned() ? (
-          <X className="size-3" />
-        ) : (
-          <PinIcon className="size-3" />
-        )}
-      </Button>
-    ),
-    size: 35,
-    enableResizing: false,
-    enableColumnFilter: false,
-  }, */
   {
-    accessorFn: (row) => `${row?.sandboxID}-${row?.clientID}`,
+    accessorKey: 'sandboxID',
     header: 'ID',
-    cell: ({ getValue }) => (
-      <div className="text-fg-500 truncate font-mono text-xs">
-        {getValue() as string}
-      </div>
-    ),
-    size: 300,
+    cell: IdCell,
+    size: 190,
     minSize: 100,
     enableColumnFilter: false,
     enableSorting: false,
@@ -157,35 +132,7 @@ export const COLUMNS: ColumnDef<SandboxWithMetrics>[] = [
     accessorKey: 'templateID',
     id: 'template',
     header: 'TEMPLATE',
-    cell: ({ getValue, table }) => {
-      const templateId = getValue() as string
-      const template: Template | undefined = table
-        .getState()
-        // @ts-expect-error - templates state not in type definition
-        .templates.find((t: Template) => t.templateID === templateId)
-
-      const { selectedTeamSlug, selectedTeamId } = useServerContext()
-
-      const router = useRouter()
-
-      if (!selectedTeamSlug || !selectedTeamId) return null
-
-      return (
-        <Button
-          variant="link"
-          className="text-fg h-auto p-0 text-xs normal-case"
-          onClick={() => {
-            useTemplateTableStore.getState().setGlobalFilter(templateId)
-            router.push(
-              PROTECTED_URLS.TEMPLATES(selectedTeamSlug ?? selectedTeamId)
-            )
-          }}
-        >
-          {template?.aliases?.[0] ?? templateId}
-          <ArrowUpRight className="size-3" />
-        </Button>
-      )
-    },
+    cell: TemplateCell,
     size: 250,
     minSize: 180,
     filterFn: 'arrIncludesSome',
@@ -193,111 +140,44 @@ export const COLUMNS: ColumnDef<SandboxWithMetrics>[] = [
   },
   {
     id: 'cpuUsage',
-    accessorFn: (row) => row.metrics?.cpuUsedPct ?? 0,
     header: 'CPU Usage',
-    cell: ({ getValue, row }) => {
-      const cpuRaw = getValue() as number
-      const cpuPercentage = Math.round(cpuRaw)
-      const hasMetrics = row.original.metrics != null
-
-      // color: <70 neutral, 70-89 warning, >=90 error
-      const textClassName = cn(
-        cpuPercentage >= 90
-          ? 'text-error'
-          : cpuPercentage >= 70
-            ? 'text-warning'
-            : 'text-fg'
-      )
-
-      return (
-        <span
-          className={cn(
-            'text-fg-500 flex w-full items-center gap-0.5 truncate font-mono whitespace-nowrap'
-          )}
-        >
-          {hasMetrics ? (
-            <span className={cn(textClassName)}>{cpuPercentage}% </span>
-          ) : (
-            <span>n/a</span>
-          )}
-          <span className="text-fg-500 mx-1">·</span>
-          <span className="text-contrast-2">{row.original.cpuCount}</span> Core
-          {row.original.cpuCount > 1 ? 's' : ''}
-        </span>
-      )
-    },
+    accessorFn: (row) => row.metrics?.cpuUsedPct,
+    cell: ({ row }) => (
+      <CpuUsageCell
+        metrics={row.original.metrics}
+        cpuCount={row.original.cpuCount ?? null}
+      />
+    ),
     size: 175,
     minSize: 120,
-    // @ts-expect-error resourceRange is not a valid filterFn
-    filterFn: 'resourceRange',
+    enableSorting: true,
+    enableColumnFilter: false,
   },
   {
     id: 'ramUsage',
-    accessorFn: (row) => {
-      if (!row.metrics?.memUsedMb || !row.metrics?.memTotalMb) return 0
-      return (row.metrics?.memUsedMb / row.metrics?.memTotalMb) * 100
-    },
     header: 'Memory Usage',
-    cell: ({ getValue, row }) => {
-      const ramRaw = getValue() as number
-      const ramPercentage = Math.round(ramRaw)
-      const hasMetrics =
-        row.original.metrics != null && row.original.metrics.memUsedMb != null
-
-      const totalRamMB = row.original.memoryMB.toLocaleString()
-
-      // color: <70 neutral, 70-94 warning, >=95 error
-      const textClassName = cn(
-        ramPercentage >= 95
-          ? 'text-error'
-          : ramPercentage >= 70
-            ? 'text-warning'
-            : 'text-fg'
-      )
-
-      const usedRamMB = hasMetrics
-        ? row.original.metrics.memUsedMb.toLocaleString()
-        : '0'
-
-      return (
-        <span
-          className={cn(
-            'text-fg-500 flex items-center gap-0.5 truncate font-mono whitespace-nowrap'
-          )}
-        >
-          {hasMetrics ? (
-            <span className={cn(textClassName)}>{ramPercentage}% </span>
-          ) : (
-            <span>n/a</span>
-          )}
-          <span className="text-fg-500 mx-1">·</span>
-          <span className={textClassName}>{usedRamMB}</span> /{' '}
-          <span className="text-contrast-1">{totalRamMB} </span> MB
-        </span>
-      )
+    accessorFn: (row) => {
+      if (row.metrics?.memUsedMb && row.metrics.memTotalMb) {
+        return (row.metrics.memUsedMb / row.metrics.memTotalMb) * 100
+      }
+      return 0
     },
+    cell: ({ row }) => (
+      <RamUsageCell
+        metrics={row.original.metrics}
+        memoryMB={row.original.memoryMB}
+      />
+    ),
     size: 175,
     minSize: 160,
-    // @ts-expect-error resourceRange is not a valid filterFn
-    filterFn: 'resourceRange',
+    enableSorting: true,
+    enableColumnFilter: false,
   },
   {
     id: 'metadata',
     accessorFn: (row) => JSON.stringify(row.metadata ?? {}),
     header: 'Metadata',
-    cell: ({ getValue }) => {
-      const value = getValue() as string
-      const json = useMemo(() => JSON.parse(value), [value])
-
-      return (
-        <JsonPopover
-          className="text-fg-500 hover:text-fg hover:underline"
-          json={json}
-        >
-          {value}
-        </JsonPopover>
-      )
-    },
+    cell: MetadataCell,
     filterFn: 'includesStringSensitive',
     enableGlobalFilter: true,
     size: 200,
@@ -307,25 +187,7 @@ export const COLUMNS: ColumnDef<SandboxWithMetrics>[] = [
     id: 'startedAt',
     accessorKey: 'startedAt',
     header: 'Started At',
-    cell: ({ row, getValue }) => {
-      const dateValue = getValue() as string
-
-      const dateTimeString = useMemo(() => {
-        return new Date(dateValue).toUTCString()
-      }, [dateValue])
-
-      const [day, date, month, year, time, timezone] = useMemo(() => {
-        return dateTimeString.split(' ')
-      }, [dateTimeString])
-
-      return (
-        <div className={cn('h-full truncate font-mono text-xs')}>
-          <span className="text-fg-500">{`${day} ${date} ${month} ${year}`}</span>{' '}
-          <span className="text-fg">{time}</span>{' '}
-          <span className="text-fg-500">{timezone}</span>
-        </div>
-      )
-    },
+    cell: StartedAtCell,
     size: 250,
     minSize: 140,
     // @ts-expect-error dateRange is not a valid filterFn
