+++ src/features/dashboard/templates/table-config.tsx
@@ -1,8 +1,5 @@
-/* eslint-disable react-hooks/rules-of-hooks */
-
 'use client'
 
-import { MoreVertical, Lock, LockOpen, Cpu } from 'lucide-react'
 import {
   ColumnDef,
   FilterFn,
@@ -13,32 +10,18 @@ import {
 } from '@tanstack/react-table'
 import { rankItem } from '@tanstack/match-sorter-utils'
 import { DefaultTemplate, Template } from '@/types/api'
-import { useToast } from '@/lib/hooks/use-toast'
-import { Button } from '@/ui/primitives/button'
-import {
-  DropdownMenu,
-  DropdownMenuContent,
-  DropdownMenuGroup,
-  DropdownMenuItem,
-  DropdownMenuLabel,
-  DropdownMenuSeparator,
-  DropdownMenuTrigger,
-} from '@/ui/primitives/dropdown-menu'
-import {
-  deleteTemplateAction,
-  updateTemplateAction,
-} from '@/server/templates/templates-actions'
-import { useMemo, useState } from 'react'
-import { Badge } from '@/ui/primitives/badge'
-import { CgSmartphoneRam } from 'react-icons/cg'
-import { useSelectedTeam } from '@/lib/hooks/use-teams'
-import { Loader } from '@/ui/loader'
-import { AlertDialog } from '@/ui/alert-dialog'
+import { useMemo } from 'react'
 import posthog from 'posthog-js'
-import { cn } from '@/lib/utils'
-import { useAction } from 'next-safe-action/hooks'
-import { defaultSuccessToast, defaultErrorToast } from '@/lib/hooks/use-toast'
-import { ByE2BBadge } from '@/features/dashboard/templates/by-e2b-badge'
+import {
+  ActionsCell,
+  TemplateIdCell,
+  TemplateNameCell,
+  CpuCell,
+  MemoryCell,
+  CreatedAtCell,
+  UpdatedAtCell,
+  VisibilityCell,
+} from './table-cells'
 
 // FILTERS
 export const fuzzyFilter: FilterFn<unknown> = (
@@ -89,151 +72,14 @@ export const useColumns = (deps: unknown[]) => {
         enableGlobalFilter: false,
         enableResizing: false,
         size: 35,
-        cell: ({ row }) => {
-          const template = row.original
-          const selectedTeam = useSelectedTeam()
-          const { toast } = useToast()
-          const [isDeleteDialogOpen, setIsDeleteDialogOpen] = useState(false)
-
-          const { execute: executeUpdateTemplate, isExecuting: isUpdating } =
-            useAction(updateTemplateAction, {
-              onSuccess: ({ input }) => {
-                toast(
-                  defaultSuccessToast(
-                    `Template is now ${input.props.Public ? 'public' : 'private'}.`
-                  )
-                )
-              },
-              onError: (error) => {
-                toast(
-                  defaultErrorToast(
-                    error.error.serverError || 'Failed to update template.'
-                  )
-                )
-              },
-            })
-
-          const { execute: executeDeleteTemplate, isExecuting: isDeleting } =
-            useAction(deleteTemplateAction, {
-              onSuccess: () => {
-                toast(defaultSuccessToast('Template has been deleted.'))
-              },
-              onError: (error) => {
-                toast(
-                  defaultErrorToast(
-                    error.error.serverError || 'Failed to delete template.'
-                  )
-                )
-              },
-              onSettled: () => {
-                setIsDeleteDialogOpen(false)
-              },
-            })
-
-          const togglePublish = async () => {
-            if (!selectedTeam) {
-              return
-            }
-
-            executeUpdateTemplate({
-              templateId: template.templateID,
-              props: {
-                Public: !template.public,
-              },
-            })
-          }
-
-          const deleteTemplate = async () => {
-            if (!selectedTeam) {
-              return
-            }
-
-            executeDeleteTemplate({
-              templateId: template.templateID,
-            })
-          }
-
-          return (
-            <>
-              <AlertDialog
-                open={isDeleteDialogOpen}
-                onOpenChange={setIsDeleteDialogOpen}
-                title="Delete Template"
-                description="Are you sure you want to delete this template? This action cannot be undone."
-                confirm="Delete"
-                onConfirm={() => deleteTemplate()}
-                confirmProps={{
-                  disabled: isDeleting,
-                  loading: isDeleting,
-                }}
-              />
-
-              <DropdownMenu>
-                <DropdownMenuTrigger asChild>
-                  <Button
-                    variant="ghost"
-                    size="icon"
-                    className="text-fg-500 size-5"
-                    disabled={
-                      isUpdating || isDeleting || 'isDefault' in template
-                    }
-                  >
-                    {isUpdating ? (
-                      <Loader className="size-4" />
-                    ) : (
-                      <MoreVertical className="size-4" />
-                    )}
-                  </Button>
-                </DropdownMenuTrigger>
-                <DropdownMenuContent align="end">
-                  <DropdownMenuGroup>
-                    <DropdownMenuLabel>General</DropdownMenuLabel>
-                    <DropdownMenuItem
-                      onClick={togglePublish}
-                      disabled={isUpdating || isDeleting}
-                    >
-                      {template.public ? (
-                        <>
-                          <Lock className="!size-3" />
-                          Set Private
-                        </>
-                      ) : (
-                        <>
-                          <LockOpen className="!size-3" />
-                          Set Public
-                        </>
-                      )}
-                    </DropdownMenuItem>
-                  </DropdownMenuGroup>
-
-                  <DropdownMenuSeparator />
-
-                  <DropdownMenuGroup>
-                    <DropdownMenuLabel>Danger Zone</DropdownMenuLabel>
-                    <DropdownMenuItem
-                      variant="error"
-                      onClick={() => setIsDeleteDialogOpen(true)}
-                      disabled={isUpdating || isDeleting}
-                    >
-                      X Delete
-                    </DropdownMenuItem>
-                  </DropdownMenuGroup>
-                </DropdownMenuContent>
-              </DropdownMenu>
-            </>
-          )
-        },
+        cell: ActionsCell,
       },
       {
         accessorKey: 'templateID',
         header: 'ID',
         size: 160,
         minSize: 120,
-        cell: ({ row }) => (
-          <div className="text-fg-500 truncate font-mono text-xs">
-            {row.getValue('templateID')}
-          </div>
-        ),
+        cell: TemplateIdCell,
       },
       {
         accessorKey: 'name',
@@ -241,37 +87,14 @@ export const useColumns = (deps: unknown[]) => {
         header: 'Name',
         size: 160,
         minSize: 120,
-        cell: ({ row, getValue }) => (
-          <div
-            className={cn(
-              'flex items-center gap-2 truncate font-mono font-medium',
-              {
-                'text-fg-500': !getValue(),
-              }
-            )}
-          >
-            <span>{(getValue() as string) ?? 'N/A'}</span>
-            {'isDefault' in row.original && row.original.isDefault && (
-              <ByE2BBadge />
-            )}
-          </div>
-        ),
+        cell: TemplateNameCell,
       },
       {
         accessorKey: 'cpuCount',
         header: 'CPU',
         size: 125,
         minSize: 125,
-        cell: ({ row }) => {
-          const cpuCount = row.getValue('cpuCount') as number
-          return (
-            <Badge className="text-fg-500 font-mono whitespace-nowrap">
-              <Cpu className="text-contrast-2 size-2.5" />{' '}
-              <span className="text-contrast-2">{cpuCount}</span> core
-              {cpuCount > 1 ? 's' : ''}
-            </Badge>
-          )
-        },
+        cell: CpuCell,
         filterFn: 'equals',
       },
       {
@@ -279,18 +102,7 @@ export const useColumns = (deps: unknown[]) => {
         header: 'Memory',
         size: 140,
         minSize: 140,
-        cell: ({ row }) => {
-          const memoryMB = row.getValue('memoryMB') as number
-          return (
-            <Badge className="text-fg-500 font-mono whitespace-nowrap">
-              <CgSmartphoneRam className="text-contrast-1 size-2.5" />{' '}
-              <span className="text-contrast-1">
-                {memoryMB.toLocaleString()}{' '}
-              </span>
-              MB
-            </Badge>
-          )
-        },
+        cell: MemoryCell,
         filterFn: 'equals',
       },
       {
@@ -300,27 +112,9 @@ export const useColumns = (deps: unknown[]) => {
         header: 'Created At',
         size: 250,
         minSize: 140,
-        cell: ({ row, getValue }) => {
-          const dateValue = getValue() as string;
-
-          const dateTimeString = useMemo(() => {
-            return new Date(dateValue).toUTCString();
-          }, [dateValue]);
-
-          const [day, date, month, year, time, timezone] = useMemo(() => {
-            return dateTimeString.split(' ');
-          }, [dateTimeString]);
-
-          return (
-            <div className={cn('h-full truncate font-mono text-xs')}>
-              <span className="text-fg-500">{`${day} ${date} ${month} ${year}`}</span>{' '}
-              <span className="text-fg">{time}</span>{' '}
-              <span className="text-fg-500">{timezone}</span>
-            </div>
-          )
-        },
+        cell: CreatedAtCell,
         sortingFn: (rowA, rowB) => {
-          return rowA.original.createdAt.localeCompare(rowB.original.createdAt);
+          return rowA.original.createdAt.localeCompare(rowB.original.createdAt)
         },
       },
       {
@@ -330,27 +124,9 @@ export const useColumns = (deps: unknown[]) => {
         size: 250,
         minSize: 140,
         enableGlobalFilter: true,
-        cell: ({ row, getValue }) => {
-          const dateValue = getValue() as string;
-
-          const dateTimeString = useMemo(() => {
-            return new Date(dateValue).toUTCString();
-          }, [dateValue]);
-
-          const [day, date, month, year, time, timezone] = useMemo(() => {
-            return dateTimeString.split(' ');
-          }, [dateTimeString]);
-
-          return (
-            <div className={cn('h-full truncate font-mono text-xs')}>
-              <span className="text-fg-500">{`${day} ${date} ${month} ${year}`}</span>{' '}
-              <span className="text-fg">{time}</span>{' '}
-              <span className="text-fg-500">{timezone}</span>
-            </div>
-          )
-        },
+        cell: UpdatedAtCell,
         sortingFn: (rowA, rowB) => {
-          return rowA.original.updatedAt.localeCompare(rowB.original.updatedAt);
+          return rowA.original.updatedAt.localeCompare(rowB.original.updatedAt)
         },
       },
       {
@@ -358,15 +134,7 @@ export const useColumns = (deps: unknown[]) => {
         header: 'Visibility',
         size: 140,
         minSize: 100,
-        cell: ({ getValue }) => (
-          <Badge
-            className={cn('text-fg-500 font-mono whitespace-nowrap', {
-              'text-success': getValue(),
-            })}
-          >
-            {getValue() ? 'Public' : 'Private'}
-          </Badge>
-        ),
+        cell: VisibilityCell,
         enableSorting: false,
         filterFn: 'equals',
       },
